Ajustes finais
